#35 mypy strict : batch 2 -27 erreurs (45 -> 18, tout dans dashboard/render)
Quick wins suite session matin :
- user_profile.py : -10 erreurs (counts dict[str, Any] init)
- bot_preferences.py : -4 erreurs (out dict[str, Any])
- portfolio_grade.py : -4 erreurs (_px Callable typed + meta fallback) +
  fix de la def conditionnelle (Callable annotation cleane sans noqa misc)
- over_cap_monitor.py : -3 erreurs (ticker = str(ln["ticker"]))
- recalib_map.py : -2 erreurs (model: Any, sklearn duck-type)
- debt_monitor.py : -1 erreur (float() cast return value)
- risk_signal_monitor.py : -1 erreur (sig_id = str(... or ""))
- thesis_invariants.py : -1 erreur (violations: list[str] annotation)
- position.py : -1 erreur (thesis_id int(_tid))

Pattern dominant : dict[str, Any] sur dicts heterogenes init.

Reste 18 erreurs toutes dans dashboard/render.py (debordement
dependances). render.py est sur le chantier #61/#66 refactor modules --
on traite ensuite. Total cumule session : 61 -> 18 (-70%).

+++ intelligence/debt_monitor.py
@@ -344,7 +344,7 @@ def _fetch_btc_drawdown180() -> float | None:
         if len(closes) < 30 or max(closes) <= 0:
             return None
         max_180 = max(closes)
-        return (closes[-1] - max_180) / max_180 * 100
+        return float((closes[-1] - max_180) / max_180 * 100)
     except Exception as e:
         log.warning(f"BTC_drawdown180 fetch: {e}")
         return None
